Some reorganization, update readme
