+++ README.md
@@ -14,47 +14,41 @@ This is a simple library that generates SVG badges, based on [badge-up](https://
 `npm install badge-up2`
 
 
-## V1 Usage
+## Basic Usage
 
 ```js
-const badge = require('badge-up2');
-const svg = badge('batman', 'component', badge.colors.green);
+const { basic, basicColors } = require('badge-up2');
+const svg = basic('batman', 'component', basicColors.green);
 ```
 
-Produces: ![example](https://cdn.rawgit.com/yahoo/badge-up/master/test/testData/good.svg)
+![example](src/test-utils/test-data/basic.svg)
 
 The color argument can be a CSS color, or one of the specially named colors
-found in `badge.colors`.
+found in `basicColors`.
 
-## V2 Usage
+## Advanced Usage
 
 With this API you can create arbitrary badges with an arbitrary number of sections.
 
-Each section is an array. The first element is the text to show, and the rest
-of the (optional) elements are attributes to apply to the text. If there are no
-attributes then a single string (instead of an array of one element) can be
-passed instead.
+Each section is either a string or a hash.
+The hash should have the following properties:
+* `text`: The text to display in the section
+* `color` (optional): The background color of the section. This can be either a six-character hex code or a css color name.
+* `strokeColor` (optional): The color of the section border
 
-The following attributes are supported:
-
-* `{hex}{hex}{hex}{hex}{hex}{hex}` six hexadecimal characters or `{css-color-name}` a named CSS color,
-    either of which is used to color the section
-* `s{{hex}{hex}{hex}{hex}{hex}{hex}}` six hexadecimal characters or `{css-color-name}` a named CSS color,
-    either of which is used to color stroke the section, e.g. `d3d3d3`
-
-The default color for the first section is `696969` and `d3d3d3` for subsequent sections.
-
-Any section can have multiple lines by putting newlines in its text.
+If `color` is not provided, then a default is used.
+The default color for the first section is `#696969` and `#d3d3d3` for subsequent sections.
+Passing a string rather than a hash is equivalent to passing a hash with only the `text` property.
 
 ```js
-const badge = require('badge-up2');
+const { badge } = require('badge-up2');
 const sections = [
     'foo/far;fun',
-    [ 'bar\nbaz', 'orange'],
-    [ 'mork "mindy"', 'olive', 's{white}'],
-    [ '<∀>', 'moccasin']
+    { text: 'bar\nbaz', color: 'orange'},
+    { text: 'mork "mindy"', color: 'olive', strokeColor: 'white'},
+    { text: '<∀>', color: 'moccasin'},
 ];
-const svg = badge.v2(sections);
+const svg = badge(sections);
 ```
 
-Produces: ![example](https://cdn.rawgit.com/yahoo/badge-up/master/test/testData/v2-example.svg)
+![example](src/test-utils/test-data/example.svg)
