+++ package.json
@@ -12,6 +12,7 @@
   "scripts": {
     "build": "tsc -p tsconfig.build.json",
     "prebuild": "rimraf dist",
+    "postbuild": "cp package.json README.md LICENSE dist",
     "test": "jest",
     "test:coverage": "jest --coverage",
     "test:watch": "jest --watch",
@@ -24,7 +25,7 @@
   },
   "homepage": "https://github.com/stevenhair/badge-up2#readme",
   "engines": {
-    "node": ">=16"
+    "node": ">=10"
   },
   "bugs": "https://github.com/stevenhair/badge-up2/issues",
   "keywords": [
@@ -50,7 +51,7 @@
     "typescript": "^5.2.2"
   },
   "dependencies": {
-    "css-color-names": "~1.0.1",
+    "css-color-names": "^1.0.1",
     "dot": "^1.1.3",
     "svgo": "^3.0.2"
   },
